requirements.txt
streamlit
pdfplumber
python-docx
pyvis
pandas
